[REFACTOR] tcp: Move connection management state into its own file
