+++ src/inetstack/protocols/layer4/tcp/established/connection_management_state.rs
@@ -1,0 +1,34 @@
+use std::net::SocketAddrV4;
+
+use crate::{
+    inetstack::{config::TcpConfig, protocols::layer4::tcp::established::ctrlblk::State},
+    runtime::network::socket::option::TcpSocketOptions,
+};
+
+/// State block representing connection management parameters in a TCP connection.
+/// This struct has only public members since these parameters must be read by all TCP
+/// modules.
+pub struct ConnectionManagementState {
+    pub local: SocketAddrV4,
+    pub remote: SocketAddrV4,
+    pub tcp_config: TcpConfig,
+    pub socket_options: TcpSocketOptions,
+    pub state: State,
+}
+
+impl ConnectionManagementState {
+    pub fn new(
+        local: SocketAddrV4,
+        remote: SocketAddrV4,
+        tcp_config: TcpConfig,
+        socket_options: TcpSocketOptions,
+    ) -> Self {
+        Self {
+            local,
+            remote,
+            tcp_config,
+            socket_options,
+            state: State::Established,
+        }
+    }
+}
