+++ src/inetstack/protocols/layer4/tcp/established/ctrlblk.rs
@@ -5,17 +5,10 @@
 // Imports
 //======================================================================================================================
 
-use crate::{
-    inetstack::{
-        config::TcpConfig,
-        protocols::layer4::tcp::established::{
-            congestion_control_state::CongestionControlState, delivery_state::DeliveryState,
-            flow_control_state::FlowControlState,
-        },
-    },
-    runtime::network::socket::option::TcpSocketOptions,
+use crate::inetstack::protocols::layer4::tcp::established::{
+    congestion_control_state::CongestionControlState, connection_management_state::ConnectionManagementState,
+    flow_control_state::FlowControlState, ordered_delivery_state::OrderedDeliveryState,
 };
-use ::std::net::SocketAddrV4;
 
 //======================================================================================================================
 // Structures
@@ -36,34 +29,6 @@ pub enum State {
     Closed,
 }
 
-/// State block representing connection management parameters in a TCP connection.
-/// This struct has only public members since these parameters must be read by all TCP
-/// modules.
-pub struct ConnectionManagementState {
-    pub local: SocketAddrV4,
-    pub remote: SocketAddrV4,
-    pub tcp_config: TcpConfig,
-    pub socket_options: TcpSocketOptions,
-    pub state: State,
-}
-
-impl ConnectionManagementState {
-    pub fn new(
-        local: SocketAddrV4,
-        remote: SocketAddrV4,
-        tcp_config: TcpConfig,
-        socket_options: TcpSocketOptions,
-    ) -> Self {
-        Self {
-            local,
-            remote,
-            tcp_config,
-            socket_options,
-            state: State::Established,
-        }
-    }
-}
-
 /// Transmission control block for representing our TCP connection.
 /// This struct has only public members because includes state for both the send and receive path and is accessed by
 /// both.
@@ -71,7 +36,7 @@ pub struct ControlBlock {
     // Connection management state, which mainly includes
     // connection constants for TCP
     pub connection_management: ConnectionManagementState,
-    pub delivery: DeliveryState,
+    pub delivery: OrderedDeliveryState,
 
     // Flow control state
     pub flow_control: FlowControlState,
@@ -87,7 +52,7 @@ pub struct ControlBlock {
 impl ControlBlock {
     pub fn new(
         connection_management: ConnectionManagementState,
-        delivery: DeliveryState,
+        delivery: OrderedDeliveryState,
         flow_control: FlowControlState,
         congestion_control: CongestionControlState,
     ) -> Self {
